Add API to manage purchase order

+++ suppliers/models/supplier.py
@@ -1,5 +1,4 @@
 import uuid
-
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
 from django.db.models import IntegerField, Max, Value
@@ -13,7 +12,7 @@ class Supplier(models.Model):
     address = models.CharField(max_length=255)
     phone = models.CharField(max_length=20)
     email = models.CharField(max_length=255, null=True, blank=True)
-    discount = models.SmallIntegerField(
+    discount = models.FloatField(
         default=0,
         validators=[MinValueValidator(0), MaxValueValidator(100)],
         null=True,
